Did some linux work during my summer vacation

+++ .config/ags/widgets/powermenu/powermenu.tsx
@@ -1,6 +1,6 @@
-import { bind, execAsync, Variable } from "astal"
+import { execAsync } from "astal"
 import { App, Astal, Gtk, Gdk } from "astal/gtk3"
-import { dialog } from "../desktop/Desktop"
+import { dialog } from "../dialog/dialog"
 
 const { TOP, BOTTOM, LEFT, RIGHT } = Astal.WindowAnchor
 const { IGNORE } = Astal.Exclusivity
@@ -8,15 +8,20 @@ const { EXCLUSIVE } = Astal.Keymode
 const { CENTER } = Gtk.Align
 
 export default (action="XYZ") => {
-    const selected = Variable<number>(1)
+    // Close window if called multiple times.
+    const window: Gtk.Window | null  = App.get_window("AstalPowermenu")
+    if (window != null) {
+        window.destroy()
+        return null
+    }
 
     const win = <window
         onKeyPressEvent={onKeyPress}
-        className={"Powermenu"}
-        name={"Powermenu"}
-        namespace={"astal-powermenu"}
+        className="AstalPowermenu"
+        name="AstalPowermenu"
+        namespace="AstalPowermenu"
         exclusivity={IGNORE}
-        keymode={Astal.Keymode.EXCLUSIVE}
+        keymode={EXCLUSIVE}
         anchor={TOP | BOTTOM | LEFT | RIGHT}>
         <box
             halign={CENTER}
@@ -26,7 +31,6 @@ export default (action="XYZ") => {
         >
             <button
                 onClicked={lock}
-                className={bind(selected).as((v) => (v === 1) ? "selected" : "")}
             >
                 <box vertical>
                     <label label={""} className={"icon"} />
@@ -35,7 +39,6 @@ export default (action="XYZ") => {
             </button>
             <button
                 onClicked={sleep}
-                className={bind(selected).as((v) => (v === 2) ? "selected" : "")}
             >
                 <box vertical>
                     <label label={""} className={"icon"} />
@@ -44,7 +47,6 @@ export default (action="XYZ") => {
             </button>
             <button
                 onClicked={logout}
-                className={bind(selected).as((v) => (v === 3) ? "selected" : "")}
             >
                 <box vertical>
                     <label label={""} className={"icon"} />
@@ -53,7 +55,6 @@ export default (action="XYZ") => {
             </button>
             <button
                 onClicked={reboot}
-                className={bind(selected).as((v) => (v === 4) ? "selected" : "")}
             >
                 <box vertical>
                     <label label={""} className={"icon"} />
@@ -62,7 +63,6 @@ export default (action="XYZ") => {
             </button>
             <button
                 onClicked={shutdown}
-                className={bind(selected).as((v) => (v === 5) ? "selected" : "")}
             >
                 <box vertical>
                     <label label={""} className={"icon"} />
@@ -83,28 +83,31 @@ export default (action="XYZ") => {
     function sleep() {
         dialog("Sleep").then((b) => {
             if (b)
-                execAsync("systemctl hybrid-sleep")            
+                execAsync("systemctl hybrid-sleep")
         })
         win.destroy()
     }
-    
+
     function logout() {
         dialog("Logout").then((b) => {
-            execAsync("swaymsg exit")        
+            if (b)
+                execAsync("swaymsg exit")
         })
         win.destroy()
     }
 
     function reboot() {
         dialog("Reboot").then((b) => {
-            execAsync("systemctl reboot")
+            if (b)
+                execAsync("systemctl reboot")
         })
         win.destroy()
     }
-    
+
     function shutdown() {
         dialog("Shutdown").then((b) => {
-            execAsync("systemctl poweroff")
+            if (b)
+                execAsync("systemctl poweroff")
         })
         win.destroy()
     }
@@ -118,23 +121,6 @@ export default (action="XYZ") => {
 
         if (keyval === Gdk.KEY_Escape) {
             close()
-        }
-
-        const currentSelectedValue = selected.get()
-
-        if (keyval === 65361 || keyval === 65056) {
-            // Shift left
-            if (currentSelectedValue <= 1)
-                selected.set(5)
-            else
-                selected.set(currentSelectedValue - 1)
-        }
-        if (keyval === 65363 || keyval == 65289) {
-            // Shift right
-            if (currentSelectedValue >= 5)
-                selected.set(1)
-            else
-                selected.set(currentSelectedValue + 1)
         }
     }
 }
